# jsonplan: Improve performance for deep objects